refactored appointment cancelation
